Add files via upload
The matrix size is specified in the files. The program first reads the matrix, then while solving the equation with the Gaussian Elimination method, it creates a vector filled with 1s in size equal to the matrix size and performs the operation. Then it reads the solution text file and compares it. Text files with sample matrix and solution values ​​used in these processes are also added.